finally

+++ src/main/java/com/choong/spr/controller/BoardController.java
@@ -1,5 +1,7 @@
 package com.choong.spr.controller;
 
+import java.security.Principal;
+import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,6 +11,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.choong.spr.domain.BoardDto;
@@ -41,8 +44,28 @@ public class BoardController {
 	}
 	
 	@PostMapping("insert")
-	public String insert(BoardDto board, RedirectAttributes rttr) {
-		boolean success = service.insertBoard(board);
+	public String insert(BoardDto board, MultipartFile[] file, 
+						 Principal principal, RedirectAttributes rttr) {
+		
+//		System.out.println(file);
+	//	System.out.println(file.getOriginalFilename());
+		//System.out.println(file.getSize());
+		
+//		if(file.getSize() > 0) {
+//			board.setFileName(file.getOriginalFilename());
+//		}
+		
+		if(file!=null) {
+			List<String> fileList = new ArrayList<String>();
+			for(MultipartFile f : file) {
+				fileList.add(f.getOriginalFilename());
+			}
+			board.setFileName(fileList);
+		}
+		
+		
+		board.setMemberId(principal.getName());
+		boolean success = service.insertBoard(board,file);
 		
 		if (success) {
 			rttr.addFlashAttribute("message", "새 글이 등록되었습니다.");
@@ -63,37 +86,58 @@ public class BoardController {
 	}
 	
 	@PostMapping("modify")
-	public String modify(BoardDto dto, RedirectAttributes rttr) {
-		boolean success = service.updateBoard(dto);
+	public String modify(BoardDto dto,
+			@RequestParam(name="removeFileList",required=false)ArrayList<String>removeFileList,
+			Principal principal, 
+			MultipartFile[] addFileList,
+			RedirectAttributes rttr) {
 		
-		if (success) {
-			rttr.addFlashAttribute("message", "글이 수정되었습니다.");
+		BoardDto oldBoard = service.getBoardById(dto.getId());
+		
+		
+		if (oldBoard.getMemberId().equals(principal.getName())) {
+			boolean success = service.updateBoard(dto,removeFileList,addFileList);
+			
+			if (success) {
+				rttr.addFlashAttribute("message", "글이 수정되었습니다.");
+			} else {
+				rttr.addFlashAttribute("message", "글이 수정되지 않았습니다.");
+			}
+			
 		} else {
-			rttr.addFlashAttribute("message", "글이 수정되지 않았습니다.");
+			rttr.addFlashAttribute("message", "권한이 없습니다.");
 		}
 		
 		rttr.addAttribute("id", dto.getId());
-		
 		return "redirect:/board/get";
 	}
 	
 	@PostMapping("remove")
-	public String remove(BoardDto dto, RedirectAttributes rttr) {
+	public String remove(BoardDto dto, Principal principal, RedirectAttributes rttr) {
 		
-		boolean success = service.deleteBoard(dto.getId());
-		
-		if (success) {
-			rttr.addFlashAttribute("message", "글이 삭제 되었습니다.");
+		// 게시물 정보 얻고
+		BoardDto oldBoard = service.getBoardById(dto.getId());
+		// 게시물 작성자(memberId)와 principal의 name과 비교해서 같을 때만 진행.
+		if (oldBoard.getMemberId().equals(principal.getName())) {
+			boolean success = service.deleteBoard(dto.getId());
+			
+			if (success) {
+				rttr.addFlashAttribute("message", "글이 삭제 되었습니다.");
+				
+			} else {
+				rttr.addFlashAttribute("message", "글이 삭제 되지않았습니다.");
+			}
 			
 		} else {
-			rttr.addFlashAttribute("message", "글이 삭제 되지않았습니다.");
+			rttr.addFlashAttribute("message", "권한이 없습니다.");
+			
+			rttr.addAttribute("id", dto.getId());
+			return "redirect:/board/get";
 		}
 		
 		return "redirect:/board/list";
 	}
-	
-	
-	
+
 	
 }
